+++ packages/javascript/lib/math/index.ts
@@ -1,5 +1,6 @@
 export { div } from './div.js';
 export { mul } from './mul.js';
 export { numPick } from './numPick.js';
+export { numUnique } from './numUnique.js';
 export { sub } from './sub.js';
 export { sum } from './sum.js';