+++ packages/javascript/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.11.2 (2026--)
 
 - **BREAKING CHANGES**: `numRandom` has renamed to `numPick`
+- `numUnique`: Add `numUnique` method
 
 ## 1.11.1 (2026-01-18)